+++ bench/upgrade_in_place.sh
@@ -112,6 +112,7 @@ pgpm.part:retiring_oid
 pgpm.part:child_oid
 pgpm.transmute_inflight:owner_pid
 pgpm.transmute_inflight:owner_backend_start
+pgpm.transmute_inflight:partition_tz
 pgpm.dropped_fk:restored_at
 pgpm.dropped_fk:validated_at
 pgpm.dropped_fk:validate_retry_after